moving the channels to its own file
